requirements.txt
AgenticBlocks.IO==0.3.0
gradio==6.11.0
pydantic==2.12.5
requests==2.33.1
litellm==1.83.3
